Moved stages to separate files from ParallelBatchImporter
